Create reports based on samples
The report-script has roughly the same functionality as the old componentlib-script in terms of reporting. The report shows samples as "groupby" categories; no other groupby-options can be chosen. An additional row shows the number of sources are used to calculate the total integrated signal of the category. Two columns are added: the RI difference (dRI) within a component, and the number of spectra that are used in the calculation of the integrated signals of the components.